Move class to product namespace.
