+++ includes/Admin/Dashboard.php
@@ -36,8 +36,8 @@ use Google\Web_Stories\Font_Post_Type;
 use Google\Web_Stories\Integrations\Site_Kit;
 use Google\Web_Stories\Locale;
 use Google\Web_Stories\Media\Types;
+use Google\Web_Stories\Product\Shopping_Vendors;
 use Google\Web_Stories\Service_Base;
-use Google\Web_Stories\Shopping_Vendors;
 use Google\Web_Stories\Story_Post_Type;
 use Google\Web_Stories\Tracking;
