+++ includes/REST_API/Products_Controller.php
@@ -28,8 +28,8 @@ namespace Google\Web_Stories\REST_API;
 
 use Google\Web_Stories\Infrastructure\HasRequirements;
 use Google\Web_Stories\Model\Product;
+use Google\Web_Stories\Product\Shopping_Vendors;
 use Google\Web_Stories\Settings;
-use Google\Web_Stories\Shopping_Vendors;
 use Google\Web_Stories\Story_Post_Type;
 use WP_Error;
 use WP_REST_Request;
